shared color test helpers and COLORLESS_THEME_FILL export
